# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditView": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChainVerification"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "events"
        ],
        "title": "AuditView",
        "type": "object"
      },
...
      "CreateExecutionRequest": {
        "additionalProperties": false,
        "description": "An execution is always created from a finished readiness analysis.",
        "properties": {
          "analysisJobId": {
            "description": "A succeeded analysis job",
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "Analysisjobid",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ExecutionMode",
            "default": "dry_run"
          },
          "startedBy": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Startedby",
            "type": "string"
          }
        },
        "required": [
          "analysisJobId",
          "startedBy"
        ],
        "title": "CreateExecutionRequest",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "The shared error shape: `{error, code, details?}`.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "Execution": {
        "additionalProperties": false,
        "properties": {
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "auditHead": {
            "default": "",
            "title": "Audithead",
            "type": "string"
          },
          "auditSeq": {
            "default": 0,
            "title": "Auditseq",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ExecutionMode"
          },
          "pauseReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pausereason"
          },
          "runbookLabel": {
            "title": "Runbooklabel",
            "type": "string"
          },
          "runbookSha256": {
            "title": "Runbooksha256",
            "type": "string"
          },
          "startedBy": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Startedby",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ExecutionState",
            "default": "CREATED"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepRun"
            },
            "title": "Steps",
            "type": "array"
          },
          "toolCallsUsed": {
            "default": 0,
            "title": "Toolcallsused",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "mode",
          "runbookLabel",
          "runbookSha256",
          "startedBy",
          "createdAt",
          "updatedAt",
          "steps"
        ],
        "title": "Execution",
        "type": "object"
      },
      "ExecutionMode": {
        "enum": [
          "dry_run",
          "live"
        ],
        "title": "ExecutionMode",
        "type": "string"
      },
...
      "ExecutionState": {
        "enum": [
          "CREATED",
          "RUNNING",
          "PAUSED",
          "COMPLETED",
          "FAILED",
          "ABORTED"
        ],
        "title": "ExecutionState",
        "type": "string"
      },
      "ExecutionSummary": {
        "additionalProperties": false,
        "properties": {
          "analysisJobId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysisjobid"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ExecutionMode"
          },
          "runbookLabel": {
            "title": "Runbooklabel",
            "type": "string"
          },
          "startedBy": {
            "title": "Startedby",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ExecutionState"
          },
          "steps": {
            "title": "Steps",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "analysisJobId",
          "state",
          "mode",
          "runbookLabel",
          "startedBy",
          "createdAt",
          "updatedAt",
          "steps"
        ],
        "title": "ExecutionSummary",
        "type": "object"
      },
...
      "JsonValue": {},
      "LifecycleRequest": {
        "additionalProperties": false,
        "description": "`pause`, `abort` and `close` need a reason; `start` and `resume` do not.",
        "properties": {
          "actor": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "actor"
        ],
        "title": "LifecycleRequest",
        "type": "object"
      },
...
      "RiskClass": {
        "enum": [
          "read",
          "write",
          "destructive"
        ],
        "title": "RiskClass",
        "type": "string"
      },
...
      "StoredExecution": {
        "additionalProperties": false,
        "description": "A past execution of an analysis, read-only: its steps and its verified audit log.",
        "properties": {
          "audit": {
            "$ref": "#/components/schemas/AuditView"
          },
          "execution": {
            "$ref": "#/components/schemas/Execution"
          }
        },
        "required": [
          "execution",
          "audit"
        ],
        "title": "StoredExecution",
        "type": "object"
      },
...
      "ToolView": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "inputSchema": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Inputschema",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "risk": {
            "$ref": "#/components/schemas/RiskClass"
          },
          "server": {
            "title": "Server",
            "type": "string"
          }
        },
        "required": [
          "server",
          "name",
          "description",
          "risk",
          "inputSchema"
        ],
        "title": "ToolView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Parse DR runbooks, check dependency health and produce readiness reports.",
    "title": "DR Readiness Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analyses/{analysis_id}/executions": {
      "get": {
        "description": "Executions created from this analysis (also readable while execution is off).",
        "operationId": "get_analysis_executions_api_v1_analyses__analysis_id__executions_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExecutionSummary"
                  },
                  "title": "Response Get Analysis Executions Api V1 Analyses  Analysis Id  Executions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Analysis Executions",
        "tags": [
          "history"
        ]
      }
    },
    "/api/v1/analyses/{analysis_id}/executions/{execution_id}": {
      "get": {
        "description": "One execution of this analysis with its steps and its verified audit log (read-only).",
        "operationId": "get_analysis_execution_api_v1_analyses__analysis_id__executions__execution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Execution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredExecution"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Analysis Execution",
        "tags": [
          "history"
        ]
      }
    },
...
    "/api/v1/executions": {
      "get": {
        "operationId": "list_executions_api_v1_executions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExecutionSummary"
                  },
                  "title": "Response List Executions Api V1 Executions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "List Executions",
        "tags": [
          "execution"
        ]
      },
      "post": {
        "description": "Plan an execution from a finished analysis; it follows the report's execution plan.\n\n`409 CONFLICT` while the analysis is unfinished or failed. Stored analyses\nwork after a restart too. `mode: live` needs `EXECUTION_ALLOW_LIVE=true`.",
        "operationId": "create_execution_api_v1_executions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Create Execution",
        "tags": [
          "execution"
        ]
      }
    },
    "/api/v1/executions/{execution_id}": {
      "get": {
        "operationId": "get_execution_detail_api_v1_executions__execution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Execution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Get Execution Detail",
        "tags": [
          "execution"
        ]
      }
    },
    "/api/v1/executions/{execution_id}/audit": {
      "get": {
        "operationId": "get_audit_api_v1_executions__execution_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "description": "Recompute the hash chain",
            "in": "query",
            "name": "verify",
            "required": false,
            "schema": {
              "default": false,
              "description": "Recompute the hash chain",
              "title": "Verify",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditView"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Get Audit",
        "tags": [
          "execution"
        ]
      }
    },
...
    "/api/v1/executions/{execution_id}/{action}": {
      "post": {
        "description": "Lifecycle. `abort` is the kill switch; calls already sent become `UNKNOWN`.",
        "operationId": "change_execution_api_v1_executions__execution_id___action__post",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "enum": [
                "start",
                "pause",
                "resume",
                "abort",
                "close"
              ],
              "title": "Action",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LifecycleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Change Execution",
        "tags": [
          "execution"
        ]
      }
    },
...
    "/api/v1/tools": {
      "get": {
        "description": "Allow-listed tools with their effective risk class (never server credentials).",
        "operationId": "list_tools_api_v1_tools_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ToolView"
                  },
                  "title": "Response List Tools Api V1 Tools Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "List Tools",
        "tags": [
          "execution"
        ]
      }
    }
  }
}